+++ daml-lf/interpreter/src/test/scala/com/digitalasset/daml/lf/speedy/EvaluationOrderTest.scala
@@ -5,7 +5,7 @@ package com.daml.lf
 package speedy
 
 import com.daml.lf.data.{FrontStack, ImmArray, Ref}
-import com.daml.lf.data.Ref.Party
+import com.daml.lf.data.Ref.{Location, Party}
 import com.daml.lf.interpretation.{Error => IE}
 import com.daml.lf.language.Ast._
 import com.daml.lf.language.LanguageVersion
@@ -19,13 +19,51 @@ import com.daml.lf.testing.parser.{ParserParameters, defaultPackageId}
 import com.daml.lf.value.Value
 import com.daml.lf.value.Value.{ValueParty, ValueRecord}
 import com.daml.logging.LoggingContext
+import com.daml.scalautil.Statement.discard
 import org.scalatest.Inside
 import org.scalatest.freespec.AnyFreeSpec
 import org.scalatest.matchers.should.Matchers
 
+import scala.collection.mutable.ArrayBuffer
 import scala.util.{Failure, Success, Try}
 
 class EvaluationOrderTest extends AnyFreeSpec with Matchers with Inside {
+
+  class TestTraceLog extends TraceLog {
+    private val messages: ArrayBuffer[(String, Option[Location])] = new ArrayBuffer()
+
+    override def add(message: String, optLocation: Option[Location])(implicit
+        loggingContext: LoggingContext
+    ): Unit = {
+      discard(messages += ((message, optLocation)))
+    }
+
+    def tracePF[X, Y](text: String, pf: PartialFunction[X, Y]): PartialFunction[X, Y] = {
+      case x if { add(text, None)(LoggingContext.ForTesting); pf.isDefinedAt(x) } => pf(x)
+    }
+
+    def traceMap[X, Y](text: String, pf: Map[X, Y]): Map[X, Y] = new Map[X, Y] {
+      override def apply(key: X): Y = {
+        add(text, None)(LoggingContext.ForTesting)
+        pf(key)
+      }
+
+      override def get(key: X): Option[Y] = {
+        add(text, None)(LoggingContext.ForTesting)
+        pf.get(key)
+      }
+
+      override def iterator: Iterator[(X, Y)] = pf.iterator
+
+      override def removed(key: X): Map[X, Y] = pf.removed(key)
+
+      override def updated[Y1 >: Y](key: X, value: Y1): Map[X, Y1] = pf.updated(key, value)
+    }
+
+    override def iterator: Iterator[(String, Option[Location])] = messages.iterator
+
+    def getMessages: Seq[String] = messages.view.map(_._1).toSeq
+  }
 
   private[this] implicit def logContext: LoggingContext = LoggingContext.ForTesting
