+++ daml-lf/interpreter/src/test/scala/com/digitalasset/daml/lf/speedy/SpeedyTestLib.scala
@@ -138,39 +138,3 @@ private[speedy] object SpeedyTestLib {
   ): PureCompiledPackages =
     typeAndCompile(Map(parserParameter.defaultPackageId -> pkg))
 }
-
-class TestTraceLog extends TraceLog {
-  private val messages: ArrayBuffer[(String, Option[Location])] = new ArrayBuffer()
-
-  override def add(message: String, optLocation: Option[Location])(implicit
-      loggingContext: LoggingContext
-  ): Unit = {
-    discard(messages += ((message, optLocation)))
-  }
-
-  def tracePF[X, Y](text: String, pf: PartialFunction[X, Y]): PartialFunction[X, Y] = {
-    case x if { add(text, None)(LoggingContext.ForTesting); pf.isDefinedAt(x) } => pf(x)
-  }
-
-  def traceMap[X, Y](text: String, pf: Map[X, Y]): Map[X, Y] = new Map[X, Y] {
-    override def apply(key: X): Y = {
-      add(text, None)(LoggingContext.ForTesting)
-      pf(key)
-    }
-
-    override def get(key: X): Option[Y] = {
-      add(text, None)(LoggingContext.ForTesting)
-      pf.get(key)
-    }
-
-    override def iterator: Iterator[(X, Y)] = pf.iterator
-
-    override def removed(key: X): Map[X, Y] = pf.removed(key)
-
-    override def updated[Y1 >: Y](key: X, value: Y1): Map[X, Y1] = pf.updated(key, value)
-  }
-
-  override def iterator: Iterator[(String, Option[Location])] = messages.iterator
-
-  def getMessages: Seq[String] = messages.view.map(_._1).toSeq
-}
